requirements.txt
django~=5.0.4
djangorestframework~=3.15.1
faker~=24.14.1
django-filter~=24.2